code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
paddlepaddle==2.5.1
paddleocr>=2.7.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
albumentations>=1.3.1
transformers
requests